Feature

+++ foodway4you-backend/src/controllers/menuController.js
@@ -1,9 +1,21 @@
 import MenuItem from '../models/MenuItem.js';
 import response from '../utils/responseHelper.js';
+// change 
+import Restaurant from '../models/Restaurant.js';
 
 export const create = async (req, res, next) => {
   try {
-    const doc = await MenuItem.create(req.body);
+     //change
+     const restaurant = await Restaurant.findById(req.body.restaurant);
+
+     if (!restaurant || 
+    restaurant.onboarding.status !== "approved" || 
+    !restaurant.isActive) {
+  return response.error(res, "Restaurant not available", 400);
+   }
+
+const doc = await MenuItem.create(req.body);
+
     response.success(res, doc, 'Created', 201);
   } catch (err) {
     next(err);
@@ -14,7 +26,17 @@ export const list = async (req, res, next) => {
   try {
     const filter = {};
     if (req.query.restaurant) filter.restaurant = req.query.restaurant;
-    const docs = await MenuItem.find(filter).sort({ createdAt: -1 });
+    if (filter.restaurant) {
+  const restaurant = await Restaurant.findById(filter.restaurant);
+
+  if (!restaurant || 
+      restaurant.onboarding.status !== "approved" || 
+      !restaurant.isActive) {
+    return response.success(res, []); 
+  }
+}
+
+const docs = await MenuItem.find(filter).sort({ createdAt: -1 });
     response.success(res, docs);
   } catch (err) {
     next(err);
@@ -23,7 +45,17 @@ export const list = async (req, res, next) => {
 
 export const update = async (req, res, next) => {
   try {
-    const doc = await MenuItem.findByIdAndUpdate(req.params.id, req.body, { new: true });
+
+const menu = await MenuItem.findById(req.params.id);
+if (!menu) return response.notFound(res);
+
+const restaurant = await Restaurant.findById(menu.restaurant);
+
+if (!restaurant || restaurant.owner.toString() !== req.user.id) {
+  return response.error(res, "Not allowed", 403);
+}
+
+const doc = await MenuItem.findByIdAndUpdate(req.params.id, req.body, { new: true });
     if (!doc) return response.notFound(res);
     response.success(res, doc, 'Updated');
   } catch (err) {
@@ -33,7 +65,16 @@ export const update = async (req, res, next) => {
 
 export const remove = async (req, res, next) => {
   try {
-    const doc = await MenuItem.findByIdAndDelete(req.params.id);
+    const menu = await MenuItem.findById(req.params.id);
+if (!menu) return response.notFound(res);
+
+const restaurant = await Restaurant.findById(menu.restaurant);
+
+if (!restaurant || restaurant.owner.toString() !== req.user.id) {
+  return response.error(res, "Not allowed", 403);
+}
+
+const doc = await MenuItem.findByIdAndDelete(req.params.id);
     if (!doc) return response.notFound(res);
     response.success(res, null, 'Deleted');
   } catch (err) {
